requirements_host.txt
boto3
requests
requests[socks]
requests-html
click
psycopg2
sh
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
simple-aws
python-dotenv
pygithub
python-dotenv
azure-mgmt-cdn
Flask-WTF
gunicorn
tldextract
azure-storage-blob
pycountry

